+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [6.0.0] - 2025-11-03
+### Added
+- \Ilex\Validation\HkidValidation\Enum\Reason::PatternError
+
+### Removed
+- \Ilex\Validation\HkidValidation\Enum\Reason::PattenError
+
 ## [5.0.1] - 2021-12-03
 ### Added
 - Support PHP 8.1
@@ -80,6 +87,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove PHP<=7.0 support
 
 ## [1.0.6] - 2016-01-04
+[6.0.0]: https://github.com/iLexN/HKID-Check-Digit/compare/5.0.1...6.0.0
 [5.0.1]: https://github.com/iLexN/HKID-Check-Digit/compare/4.1.0...5.0.1
 [4.1.0]: https://github.com/iLexN/HKID-Check-Digit/compare/4.0.0...4.1.0
 [4.0.0]: https://github.com/iLexN/HKID-Check-Digit/compare/3.1.2...4.0.0